Made heuristics back to int, and updated emulator logic

+++ GopSolverLib/Solver.h
@@ -24,7 +24,7 @@ public:
 		}
 	}
 
-	double getHeuristicCost() const;
+	int getHeuristicCost() const;
 	std::deque<std::pair<GameState, GameAction>> getPath() const;
 	std::string getActions() const;
